% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{abc}
\alias{abc}
\alias{alphabetise}
\alias{alphabetize}
\title{sort anything alphabetically}
\usage{
abc(text, format_output = TRUE)

alphabetise(text, format_output = TRUE)

alphabetize(text, format_output = TRUE)
}
\arguments{
\item{text}{A character vector with multiple entries or with comma separated text}

\item{format_output}{TRUE returns a character vector, FALSE returns ordered elements, inline returns text}
}
\value{
A character vector.
}
\description{
sort anything alphabetically
}
\examples{
 x <- "charlie,delta,alpha,bravo"
 abc(x, format_output = TRUE)
 abc(x, format_output = FALSE)
 abc(x, format_output = "inline")

}
