% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date.R
\name{add_date}
\alias{add_date}
\title{add_date wrapper}
\usage{
add_date(x, sep = "_")
}
\arguments{
\item{x}{a character string often a (file) name}

\item{sep}{a character string, can be set to empty}
}
\value{
a character string with today's date
}
\description{
add_date wrapper
}
\examples{
add_date("file.R", "-")
}
