% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazysave.R
\name{lazysave}
\alias{lazysave}
\title{lazysave}
\usage{
lazysave(
  x,
  dir = tempdir(),
  date = FALSE,
  datesep = "_",
  extension = "rds",
  quiet = FALSE
)
}
\arguments{
\item{x}{an object, usually a dataframe to be saved}

\item{dir}{path where you want to save}

\item{date}{TRUE appends the date, similar to \code{add_date()}}

\item{datesep}{a character string, can be set to empty, between name and date}

\item{extension}{either ".rds" (default) or ".Rdata"}

\item{quiet}{TRUE doesn't show success message}
}
\value{
a file saved on disk
}
\description{
lazysave
}
\examples{
lazysave(df, date=TRUE, datesep = "")
# More detailed examples
# create a dataframe called data 
data <- data.frame(x = 1, 
y = 1:10, 
char = sample(LETTERS[1:3], 10, replace = TRUE)) 
# create temp folder with a subfolder
# define your path
temp <- file.path(tempdir(), "demo")
# remove everything in temp
unlink(temp, recursive = TRUE)   
# create directory 
dir.create(temp)


lazysave(data)
list.files(temp) # still empty, as default path out is tempdir.

# you may append the date to the filename
lazysave(data, dir = temp, date=TRUE, datesep = "", extension = "rdata")
# return file names 
list.files(temp) # file saved.

# lazysave() doesn't mind dot(s) in extensions  (but rds is the default)
lazysave(data, dir = temp, date=TRUE, datesep = "-", extension = ".r.ds.")
# lazysave() doesn't mind capitalisation or not in extension
lazysave(data, dir = temp, extension = "RDATA")
}
