% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_unique.R
\name{summ_unique}
\alias{summ_unique}
\alias{summarise_unique}
\title{Summarise and report number of unique values and NAs}
\usage{
summ_unique(df, has.na = FALSE)

summarise_unique(df, has.na = FALSE)
}
\arguments{
\item{df}{a dataframe to summarise.}

\item{has.na}{see for each variable if it "has NA(s)" or "has no NAs".
Defaults to FALSE, which means, reporting if a variable has (any) NAs.}
}
\value{
a dataframe reporting each variable, its unique values and NAs
}
\description{
Summarise and report number of unique values as well as the presence of NAs
for each variable in a dataframe.
}
\examples{
# generate an example dataframe
df_no_na <- data.frame(x = 1, # 1 unique value
y = 1:10, # 10 unique values 
char = sample(LETTERS[1:3], 10, replace = TRUE)) # max 3 unique values
summ_unique(df_no_na) 
# add row with NAs
row <- c(NA, 10, NA)
df_with_NAs <- rbind.data.frame(df_no_na, row)
summ_unique(df_with_NAs)
summ_unique(df_with_NAs, has.na = TRUE)  # show if cols have NAs

}
