% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\alias{tempex}
\alias{temprm}
\title{Remove objects whose names match a pattern}
\usage{
cleanup(
  pattern = "(_temp|temp_)",
  envir = .GlobalEnv,
  keep.functions = TRUE,
  verbose = TRUE
)

tempex(
  pattern = "(_temp|temp_)",
  envir = .GlobalEnv,
  keep.functions = TRUE,
  verbose = TRUE
)

temprm(
  pattern = "(_temp|temp_)",
  envir = .GlobalEnv,
  keep.functions = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{pattern}{What shall be searched for.}

\item{envir}{usually the GlobalEnvironment}

\item{keep.functions}{TRUE if you do not want to remove temporary functions}

\item{verbose}{gives feedback or none}
}
\value{
empty character string
}
\description{
To remove temporary objects (usually named "temp_" or "_temp")
}
\examples{
cleanup()
# More detailed examples
## Create some noise. 
### create things to be kept
temptation_to_keep <- data.frame(x = 1, 
                         y = 1:10, 
                         char = sample(LETTERS[1:3], 10, replace = TRUE))
url_to_keep <- "https://www.nothing_temp.csv"
### create things to be deleted: 
list_temp <- c(char = sample(LETTERS[1:3], 10, replace = TRUE))
temp_url <- "https://raw.githubusercontent.com/tidyverse/dplyr/main/data-raw/starwars.csv"
temp_dir <- file.path(tempdir(), "demo")
#### define a temp function
temp_function <- function(x) {
x + 1
}
## Some noice was created.
# use function but turn off  verbose
cleanup(verbose = FALSE)

# use function with own pattern
cleanup(pattern = "_to_keep")

## create another temporary list again
temporary_list <- c("something", "else")
# use function with own pattern
cleanup(pattern="temporary_")

# use function when nothing is available that matches the pattern
cleanup(pattern="precious")

# use function when nothing is available matching the pattern - but a function
cleanup()

# remove the temp-function
cleanup(keep.functions = FALSE)
}
