% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazysave.R
\name{lazysave}
\alias{lazysave}
\title{lazysave}
\usage{
lazysave(
  x,
  dir = tempdir(),
  date = FALSE,
  datesep = "_",
  extension = "rds",
  quiet = FALSE
)
}
\arguments{
\item{x}{an object, usually a dataframe to be saved}

\item{dir}{path where you want to save}

\item{date}{TRUE appends the date, similar to \code{add_date()}}

\item{datesep}{a character string, can be set to empty, between name and date}

\item{extension}{either ".rds" (default) or ".Rdata"}

\item{quiet}{TRUE doesn't show success message}
}
\value{
a file saved on disk
}
\description{
lazysave
}
\examples{
lazysave(df, date=TRUE, datesep = "")
}
