% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcelines.R
\name{sourcelines}
\alias{sourcelines}
\title{sourcelines.R
Source specific lines in an R file (from https://gist.github.com/christophergandrud/1eb4e095974204b12af9#file-source_lines-r)}
\usage{
sourcelines(file, lines)
}
\arguments{
\item{file}{character string with the path to the file to source.}

\item{lines}{numeric vector of lines to source in \code{file}.}
}
\description{
sourcelines.R
Source specific lines in an R file (from https://gist.github.com/christophergandrud/1eb4e095974204b12af9#file-source_lines-r)
}
